refactor(tbtc): split node into per-concern files
The pkg/tbtc/node.go file had grown to ~1450 LOC covering the node
struct plus signing/coordination/inactivity executors, wallet-action
proposal handlers, the coordination layer, wallet closure, and block
helpers. Split it into per-concern files within the same package:

- node.go              core struct, constructor, accessors, DKG entry
- node_executors.go    signing/coordination/inactivity executors
- node_proposals.go    wallet-action proposal handlers
- node_coordination.go coordination layer loop
- node_wallet_closure.go closed-wallet archival and closure
- node_block.go        block-height wait helpers

Pure relocation: declarations are moved verbatim, no behavioral change.
