+++ pkg/tbtc/node_block.go
@@ -1,0 +1,58 @@
+package tbtc
+
+import (
+	"context"
+)
+
+// waitForBlockFn represents a function blocking the execution until the given
+// block height.
+type waitForBlockFn func(context.Context, uint64) error
+
+// getCurrentBlockFn represents a function returning the current block height.
+type getCurrentBlockFn func() (uint64, error)
+
+// TODO: this should become a part of BlockHeightWaiter interface.
+func (n *node) waitForBlockHeight(ctx context.Context, blockHeight uint64) error {
+	blockCounter, err := n.chain.BlockCounter()
+	if err != nil {
+		return err
+	}
+
+	wait, err := blockCounter.BlockHeightWaiter(blockHeight)
+	if err != nil {
+		return err
+	}
+
+	select {
+	case <-wait:
+	case <-ctx.Done():
+	}
+
+	return nil
+}
+
+// withCancelOnBlock returns a copy of the given ctx that is automatically
+// cancelled on the given block or when the parent ctx is done. Note that the
+// context can be cancelled earlier if the waitForBlockFn returns an error.
+func withCancelOnBlock(
+	ctx context.Context,
+	block uint64,
+	waitForBlockFn waitForBlockFn,
+) (context.Context, context.CancelFunc) {
+	blockCtx, cancelBlockCtx := context.WithCancel(ctx)
+
+	go func() {
+		defer cancelBlockCtx()
+
+		err := waitForBlockFn(ctx, block)
+		if err != nil {
+			logger.Errorf(
+				"failed to wait for block [%v]; "+
+					"context cancelled earlier than expected",
+				err,
+			)
+		}
+	}()
+
+	return blockCtx, cancelBlockCtx
+}
